Move Header and Footer out to components
